+++ PgRoutiner/DataAccess/Models/PgRoutineGroup.cs
@@ -8,6 +8,7 @@ public record PgRoutineGroup
     public string RoutineName { get; set; }
     public string Description { get; set; }
     public string Definition { get; set; }
+    public string FullDefinition { get; set; }
     public string Language { get; set; }
     public string RoutineType { get; set; }
     public string TypeUdtName { get; set; }